Major QMAKE ports!!
1. Added project/subprojects for http_parser AND libuv
2. Updates to node-native/http project files
2. Minor fix missed in previous update

+++ lib/http/src/http.cc
@@ -100,7 +100,7 @@ http::response::~response()
 {
 }
 
-void http::response::write(unsigned int length, const char* body)
+void http::response::write(size_t length, const char* body)
 {
   if(!is_response_written_)
   {
